Removed duplicate initialization check based on selector
Getting the selector used to initialize IconCaptcha in jQuery 3 is no longer possible. This caused the code that works on previous versions of jQuery to break in this version.

Instead of relying on the selector to perform the duplicate init check, the content of the widget will be checked to ensure no previous initialized captcha is present.